fix: Change yjs_clock column from integer to bigint
JavaScript Date.now() returns milliseconds since Unix epoch, which
exceeds PostgreSQL's integer max value (2147483647) after year 2038.

Error: 'value "1761142890720" is out of range for type integer'

Fix: Changed yjs_clock to BIGINT which can handle values up to
9223372036854775807.

This fixes the storage errors in Hocuspocus onStoreDocument hook.

+++ supabase/migrations/20251022_add_yjs_support.sql
@@ -4,7 +4,7 @@
 -- Add Y.js state columns
 ALTER TABLE public.trips
 ADD COLUMN IF NOT EXISTS yjs_state BYTEA,
-ADD COLUMN IF NOT EXISTS yjs_clock INTEGER DEFAULT 0;
+ADD COLUMN IF NOT EXISTS yjs_clock BIGINT DEFAULT 0;
 
 -- Drop old itinerary_document column (no backward compatibility needed)
 ALTER TABLE public.trips
